fix: undefined name errors in beeswarm
And also temporarily remove the SHAP interaction beeswarm plot code
(it's not runnable at the moment, because we ensured that the ndim of
SHAP values was 2 above, so it can never be 3 to trigger the interaction
beeswarm plot here).

The removal of code here is mostly just to appease ruff.
Will create a separate issue for this to re-introduce SHAP interaction
values plotting again later on. Maybe it should be implemented as a
separate function.